requirements.txt
black
channels==4.0.0
daphne==4.0.0
django-cors-headers==3.14.0
Django<5
djangorestframework-simplejwt<6
djangorestframework<4
psycopg[binary,pool]==3.1.9
pynacl==1.5.0
python-decouple==3.8
requests==2.28.2
the-python-bay==1.2.3
whitenoise==6.4.0
pytest-django==4.5.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1